requirements.txt
# SAE Multilingual Steering - Research Grade Dependencies
# Install: pip install -r requirements.txt

# Core ML
torch>=2.0.0
transformers>=4.40.0
datasets>=2.14.0
accelerate>=0.27.0

# Sparse Autoencoders
sae-lens>=3.0.0

# LLM Judge (Gemini API)
# NOTE: the evaluation pipeline imports `google.generativeai`, which is
# provided by `google-generativeai` (not `google-genai`).
google-generativeai>=0.7.0

# Optional: Other LLM judges
# anthropic>=0.21.0
# openai>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning (required for exp22, exp24 classifiers)
scikit-learn>=1.3.0

# Semantic Similarity (required for semantic metrics in evaluation)
sentence-transformers>=2.6.0

# Progress bars
tqdm>=4.65.0

# Flash Attention (optional, for speed)
# flash-attn>=2.0.0  # Requires CUDA

# Development
ipython>=8.0.0
jupyter>=1.0.0

# Utilities
python-dotenv>=1.0.0

# Optional lightweight language identification for Latin-script controls.
# Recommended: fastText-based LID (no compilation on many platforms).
# If installed, set LID_BACKEND=fasttext (optionally run scripts/download_fasttext_lid.py).
fast-langdetect>=0.3.0
#
# Fallback: langid (pure Python, but weaker on short text).
# If installed, set LID_BACKEND=langid.
# langid>=1.1.6
